fix: chats
