+++ src/components/index.ts
@@ -20,5 +20,4 @@ export { FormProfile } from "./form-profile";
 export { FormPasswordProfile } from "./form-password-profile";
 export { Input } from "./input";
 export { Button } from "./button";
-export { PageSelectedChats } from "../pages/SelectedChats";
 export { ErrorInputText } from "./error-input-text";
