--- src/pages/Chat/index.hbs
@@ -1,24 +1,0 @@
-{{#> PageListChat}}
-<div class="container-chat">
-        <div class="container-chat__profile">
-            {{> CardUser name="Вадим" photo=photo_user}}
-            <div class="container-chat__profile__menu">
-                <img src="{{menu}}" alt="Иконка"/>
-            </div>
-        </div>
-        <div class="container-chat__element">
-            {{> Chat chats=(listMessage)}}
-        </div>
-        <div class="container-chat__input">
-            <div class="container-chat__input__attacher">
-                <img src="{{attacher}}" alt="Иконка"/>
-            </div>
-            <div class="container-chat__input__element">
-                {{> InputMessage placeholder="Сообщение"}}
-            </div>
-            <div class="container-chat__input__button">
-                    <img src="{{arrow_circle}}" alt="Иконка">
-            </div>
-        </div>
-    </div>
-{{/PageListChat}}
