--- src/pages/Chat/index.ts
@@ -1,4 +1,0 @@
-import Handlebars from "handlebars";
-import "./style.css";
-// export { default as PageChat } from "./index.hbs?raw";
-export { PageChat } from "./chat";
