--- src/pages/Chat/style.css
@@ -1,53 +1,0 @@
-.container-chat {
-    display: flex;
-    flex-direction: column;
-    padding: 10px 20px;
-    height: 100%;
-}
-
-.container-chat__profile {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-
-    padding-bottom: 10px;
-}
-
-.container-chat__profile__menu {
-    cursor: pointer;
-}
-
-.container-chat__element {
-    display: flex;
-    flex-direction: column;
-    flex: 1 0;
-    height: 80%;
-    border-top: 1px solid var(--primary-border);
-    border-bottom: 1px solid var(--primary-border);
-    padding-top: 49px;
-    padding-bottom: 30px;
-}
-
-.container-chat__input {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-
-    width: 100%;
-    margin-top: 20px;
-}
-
-.container-chat__input__element {
-    flex: 1 0;
-}
-
-.container-chat__input__element input {
-    width: 100%;
-}
-
-
-.container-chat__input__button {
-    border: none;
-    background: var(--primary-background);
-    cursor: pointer;
-}
