+++ src/pages/Chats/index.hbs
