+++ src/pages/Chats/index.ts
@@ -1,6 +1,6 @@
 import "./style.css";
 // import PageSelectedChats from "./index.hbs?raw";
 // export { PageSelectedChats };
-export { PageSelectedChats } from "./selectedChats";
+export { PageChats } from "./selectedChats";
 
 // Handlebars.registerPartial("PageListChat", PageSelectedChats);
