+++ src/pages/Chats/selectedChats.ts
@@ -1,20 +1,68 @@
 import { Block, Props } from "../../core/Block";
 import avatar from "../../assets/photoUser.png";
 import arrow from "../../assets/arrow.svg";
+import menu from "../../assets/menu.svg";
+import photo1 from "../../assets/test_photo.jpg";
+import photo2 from "../../assets/test_photo2.png";
+import { InputMessage } from "../../components/input-message";
 
-export class PageSelectedChats extends Block {
+export class PageChats extends Block {
   constructor(props: Props) {
     super({
       ...props,
       arrow,
       search: "",
+      menu,
+      activeChath: false,
+      activeChatId: null,
+      test: () => {
+        console.log("vvvv");
+      },
       searchChat: (event: Event) => this.searchChat(event),
+      sendMessage: () => this.sendMessage(),
+      clickCard: (event: Event, id: string) => {
+        this.props.activeChath = true;
+        this.props.activeChatId = id;
+      },
       filterListChat: [],
+      listMessage: [
+        {
+          day: "19 июня",
+          messages: [
+            {
+              type: "text",
+              value:
+                "Привет! Смотри, тут всплыл интересный кусок лунной космической истории — НАСА в какой-то момент попросила Хассельблад адаптировать модель SWC для полетов на Луну. Сейчас мы все знаем что астронавты летали с моделью 500 EL — и к слову говоря, все тушки этих камер все еще находятся на поверхности Луны, так как астронавты с собой забрали только кассеты с пленкой. Хассельблад в итоге адаптировал SWC для космоса, но что-то пошло не так и на ракету они так никогда и не попали. Всего их было произведено 25 штук, одну из них недавно продали на аукционе за 45000 евро.",
+              myMessage: false,
+              time: "10:30",
+            },
+            {
+              type: "text",
+              value: "asd",
+              myMessage: true,
+              time: "10:30",
+            },
+            {
+              type: "photo",
+              value: photo1,
+              myMessage: false,
+              time: "10:30",
+            },
+            {
+              type: "photo",
+              value: photo2,
+              myMessage: true,
+              time: "10:31",
+            },
+          ],
+        },
+      ],
     });
   }
 
   listChat: Array<Record<string, string>> = [
     {
+      id: "1",
       name: "Андрей",
       message: "Изображение",
       photo: avatar,
@@ -22,6 +70,7 @@ export class PageSelectedChats extends Block {
       count: "2",
     },
     {
+      id: "2",
       name: "Киноклуб",
       meMessage: "стикер",
       photo: avatar,
@@ -29,6 +78,7 @@ export class PageSelectedChats extends Block {
       count: "",
     },
     {
+      id: "3",
       name: "Илья",
       message:
         "Друзья, у меня для вас особенный выпуск новостей! Друзья, у меня для вас особенный выпуск новостей! Друзья, у меня для вас особенный выпуск новостей!",
@@ -36,8 +86,9 @@ export class PageSelectedChats extends Block {
       time: "15:12",
       count: "4",
     },
-    { name: "Вадим", message: "Круто!", photo: avatar, time: "15:12" },
+    { id: "4", name: "Вадим", message: "Круто!", photo: avatar, time: "15:12" },
     {
+      id: "5",
       name: "тет-а-теты",
       message:
         "И Human Interface Guidelines и Material Design рекомендуют И Human Interface Guidelines и Material Design рекомендуют",
@@ -45,6 +96,7 @@ export class PageSelectedChats extends Block {
       time: "Пт",
     },
     {
+      id: "6",
       name: "Design Destroyer",
       message:
         "Миллионы россиян ежедневно проводят десятки часов свое Миллионы россиян ежедневно проводят десятки часов свое",
@@ -52,6 +104,7 @@ export class PageSelectedChats extends Block {
       time: "Ср",
     },
     {
+      id: "7",
       name: "Стас Рогозин",
       message:
         "В 2008 году художник Jon Rafman  начал собирать В 2008 году художник Jon Rafman  начал собирать",
@@ -59,6 +112,7 @@ export class PageSelectedChats extends Block {
       time: "Пн",
     },
     {
+      id: "8",
       name: "Петров",
       message:
         "Так увлёкся работой по курсу, что совсем забыл его анонсир Так увлёкся работой по курсу, что совсем забыл его анонсир",
@@ -66,12 +120,14 @@ export class PageSelectedChats extends Block {
       time: "Пн",
     },
     {
+      id: "9",
       name: "Настя",
       message: "Можно или сегодня или завтра вечером.",
       photo: avatar,
       time: "1 Мая 2020",
     },
     {
+      id: "10",
       name: "Design Destroyer",
       message:
         "Миллионы россиян ежедневно проводят десятки часов свое Миллионы россиян ежедневно проводят десятки часов свое",
@@ -79,6 +135,7 @@ export class PageSelectedChats extends Block {
       time: "Ср",
     },
     {
+      id: "11",
       name: "Стас Рогозин",
       message:
         "В 2008 году художник Jon Rafman  начал собирать В 2008 году художник Jon Rafman  начал собирать",
@@ -86,6 +143,7 @@ export class PageSelectedChats extends Block {
       time: "Пн",
     },
     {
+      id: "12",
       name: "Петров",
       message:
         "Так увлёкся работой по курсу, что совсем забыл его анонсир Так увлёкся работой по курсу, что совсем забыл его анонсир",
@@ -93,6 +151,7 @@ export class PageSelectedChats extends Block {
       time: "Пн",
     },
     {
+      id: "13",
       name: "Настя",
       message: "Можно или сегодня или завтра вечером.",
       photo: avatar,
@@ -116,6 +175,13 @@ export class PageSelectedChats extends Block {
     });
   }
 
+  sendMessage() {
+    const component = this.refs.inputMessage as InputMessage;
+    const message = component.getValue();
+    if (!message) return;
+    console.log(message);
+  }
+
   protected render() {
     return `<div class="wrapper-chat">
     <div class="wrapper-choice">
@@ -134,20 +200,45 @@ export class PageSelectedChats extends Block {
             <ul class="list-chat">
                 <div class="list-chat__delimiter"></div>
                 {{#each filterListChat}}
-                <li class="list-chat__element">
-                    {{{ CardUser photo=this.photo name=this.name message=this.message time=this.time count=this.count meMessage=this.meMessage}}}
-                </li>
-                <div class="list-chat__delimiter"></div>
+                  <li class="list-chat__element {{#if (isEqual id ../this.activeChatId) }}active{{/if}}">
+                      {{{ CardUser onClick=../this.clickCard id=id photo=this.photo name=this.name message=this.message time=this.time 
+                        count=this.count meMessage=this.meMessage }}}
+                  </li>
+                  <div class="list-chat__delimiter"></div>
                 {{/each}}
             </ul>
         </div>
     </div>
-    {{#if (isEqual component false)}}
+    {{#if (isEqual activeChath false)}}
         <div class="empty-chat">
             <h3>Выберите чат чтобы отправить сообщение</h3>
         </div>
+    {{else}}
+      <div class="container-chat">
+        <div class="container-chat__profile">
+              {{{ CardUser name="Вадим" photo=photo_user }}}
+              <div class="container-chat__profile__menu">
+                  <img src="{{menu}}" alt="Иконка"/>
+              </div>
+        </div>
+          <div class="container-chat__element">
+              {{{ Chat listMessage=listMessage }}}
+          </div>
+          <div class="container-chat__input">
+              <div class="container-chat__input__attacher">
+                  <img src="{{attacher}}" alt="Иконка"/>
+              </div>
+              <div class="container-chat__input__element">
+                  {{{ InputMessage ref="inputMessage" placeholder="Сообщение" }}}
+              </div>
+              {{#Button class="container-chat__input__button" onClick=sendMessage}}
+                <img src="{{arrow_circle}}" alt="Иконка">
+              {{/Button}}
+              <div class="container-chat__input__button">
+              </div>
+      </div>
+      </div>
     {{/if}}
-</div>
-`;
+</div>`;
   }
 }
