+++ src/pages/Nav/index.hbs
@@ -1,8 +1,7 @@
 <nav class="container-nav">
   <a page="login">Вход</a>
   <a page="signin">Регистрация</a>
-  <a page="listChat">Выбрать чат</a>
-  <a page="chat">Чат</a>
+  <a page="chats">Чат</a>
   <a page="profile">Профиль</a>
   <a page="modals">Модальные окна</a>
   <a page="error404">Ошибка 404</a>
