+++ src/pages/index.ts
@@ -1,8 +1,7 @@
 export { NavPage } from "./Nav";
 export { LoginPage } from "./Login";
 export { PageSign } from "./Signin";
-export { PageSelectedChats } from "./SelectedChats";
-export { PageChat } from "./Chat";
+export { PageChats } from "./Chats";
 export { PageProfile } from "./Profile";
 export { PageModals } from "./modals";
 export { PageError } from "./error";
